+++ cmd/root.go
@@ -29,10 +29,10 @@ func init() {
 func initConfig() {
 	// Find config directory.
 	configDir, err := os.UserConfigDir()
+	cobra.CheckErr(err)
 	configDir += string(os.PathSeparator) + "clockify2cats"
 
 	os.MkdirAll(configDir, os.ModePerm)
-	cobra.CheckErr(err)
 
 	// Search config in home directory with name ".clockify2cats" (without extension).
 	viper.AddConfigPath(configDir)